HU_Project/requirements.txt
asgiref==3.5
Django==4.0.3
djangorestframework==3.13.1
psycop2==2.9.3
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1